organize routing domain
