--- packages/houdini-react/src/runtime/index.ts
@@ -1,4 +1,0 @@
-export * from './hooks'
-export { HoudiniProvider } from './context'
-export * from './routing'
-export * from './Router'
