+++ packages/houdini-react/src/runtime/index.tsx
@@ -4,7 +4,11 @@ import React from 'react'
 
 import client from './client'
 import manifest from './manifest'
-import { Router as RouterImpl } from './routing/router'
+import { Router as RouterImpl } from './routing'
+
+export * from './hooks'
+export { HoudiniProvider } from './context'
+export * from './routing'
 
 export function Router({ cache = getCache() }: { cache: Cache }) {
 	return <RouterImpl cache={cache} manifest={manifest} client={client} />
