+++ packages/houdini-react/src/runtime/routing/components/Link.tsx
@@ -1,6 +1,6 @@
 import React, { useTransition } from 'react'
 
-import { useRouterContext } from './router'
+import { useRouterContext } from './Router'
 
 export function Link({
 	href,
