+++ packages/houdini-react/src/runtime/routing/components/index.ts
@@ -1,0 +1,2 @@
+export { Link } from './Link'
+export { Router, type RouterManifest } from './Router'
