+++ packages/houdini-react/src/runtime/routing/index.ts
@@ -1,2 +1,1 @@
-export { Link } from './link'
-export type { RouterManifest } from './router'
+export * from './components'
